+++ src/main/java/lotto/Main.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.LottoService;
+
 public class Main {
     public static void main(String[] args) {
         LottoService lottoService = new LottoService();
