+++ src/main/java/lotto/domain/model/LottoRank.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.model;
 
 import java.util.List;
 import java.util.Map;
