+++ src/main/java/lotto/domain/model/LottoSeller.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.domain.model;
+
+import lotto.view.model.UserMoneyInput;
 
 public class LottoSeller {
     public static final int PRICE = 1000;
