+++ src/main/java/lotto/domain/model/Money.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.model;
 
 public class Money {
     private final int amount;
