+++ src/main/java/lotto/view/model/UserMoneyInput.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.view.model;
+
+import lotto.domain.model.Money;
 
 public class UserMoneyInput {
     private Money money;
