+++ src/test/java/lotto/LottoResultTest.java
@@ -1,9 +1,12 @@
 package lotto;
 
+import lotto.domain.model.Lotto;
+import lotto.domain.model.LottoRank;
+import lotto.domain.model.LottoResult;
+import lotto.domain.model.LottoSeller;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
