+++ src/test/java/lotto/LottoSellerTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.model.LottoSeller;
+import lotto.view.model.UserMoneyInput;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
